+++ src/TemperatureDataBaseWPF/Services/SerialDataAcquisition.cs
@@ -1,0 +1,12 @@
+﻿using System.IO.Ports;
+
+namespace TemperatureDataBaseWPF.Services
+{
+    public class SerialDataAcquisition
+    {
+        public SerialDataAcquisition()
+        {
+            
+        }
+    }
+}